update for RN 0.42.0 and RNFS 2.1.0
