+++ example/ios/pdfexample/Images.xcassets/AppIcon.appiconset/Contents.json
